requirements.txt
numpy<2.0
pandas>=1.5.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
requests>=2.28.0
python-dotenv>=0.19.0
pyyaml>=6.0
toml>=0.10.2
vaderSentiment>=3.3.2
ib_insync>=0.9.86
textblob>=0.17.1
langchain>=0.1.0
langgraph>=0.0.20
newsapi-python>=0.2.7
shap>=0.41.0
neuralforecast>=1.6.0
statsforecast>=1.5.0
scipy>=1.10.0
statsmodels>=0.14.0
pmdarima>=2.0.0
# GPU-enabled versions handled in Dockerfile with special pip indices
openai>=1.0.0
# Interactive Analyst Mode dependencies
streamlit>=1.28.0
plotly>=5.15.0
seaborn>=0.12.0
